api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.3
sqlalchemy==2.0.30
asyncpg==0.29.0
redis==5.0.7
celery==5.3.6
python-multipart==0.0.9
httpx==0.27.0
orjson==3.10.4
prometheus-client==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
